+++ linphone-app/src/components/other/colors/ColorListModel.hpp
@@ -126,6 +126,10 @@ class ColorListModel : public ProxyListModel {
 		ADD_COLOR("telkeypad_bg", "#4D5B66", "Background for phone keypad")
 		ADD_COLOR("telkeypad_fg", "#E4E4E4", "Foreground for phone keypad")
 		ADD_COLOR("telkeypad_h", "#B1B1B1", "Foreground for phone keypad")
+		
+		ADD_COLOR("progress_bg", "black", "Background of round progress bar")
+		ADD_COLOR("progress_remaining_fg", "white", "Remaining progression color")
+				
 
 // Keywords: 'mKeywordsMap'
 //		s=standard, ma=main, l=list, sc=screen, me=menu