+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to set the display name in "using an account" tab of assistant.
 - Long pressed buttons.
 - Phone dialpad on main window.
+- Animated file in chats/notifications.
+- Round progress bar for transferring a file and allow to cancel it.
 
 ### Fixed
 - Crash on exit.